+++ src/Container/League/Container.php
@@ -1,18 +1,17 @@
 <?php
 
-namespace Rougin\Slytherin\IoC\League;
+namespace Rougin\Slytherin\Container\League;
 
 /**
  * LeagueContainer
  *
  * A simple implementation of a container that is based on League\Container.
- * NOTE: To be removed in v1.0.0
  *
  * http://container.thephpleague.com
  *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class Container extends \Rougin\Slytherin\Container\League\Container
+class Container extends \League\Container\Container
 {
 }