+++ src/Dispatching/DispatcherInterface.php
@@ -6,18 +6,11 @@ namespace Rougin\Slytherin\Dispatching;
  * Dispatcher Interface
  *
  * An interface for handling third party route dispatchers.
+ * NOTE: To be removed in v1.0.0
  *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-interface DispatcherInterface
+interface DispatcherInterface extends \Rougin\Slytherin\Routing\DispatcherInterface
 {
-    /**
-     * Dispatches against the provided HTTP method verb and URI.
-     *
-     * @param  string $httpMethod
-     * @param  string $uri
-     * @return array
-     */
-    public function dispatch($httpMethod, $uri);
 }