+++ src/Dispatching/RouterInterface.php
@@ -6,25 +6,11 @@ namespace Rougin\Slytherin\Dispatching;
  * Router Interface
  *
  * An interface for handling third party routers.
+ * NOTE: To be removed in v1.0.0
  *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-interface RouterInterface
+interface RouterInterface extends \Rougin\Slytherin\Routing\RouterInterface
 {
-    /**
-     * Adds a new route.
-     *
-     * @param string|string[] $httpMethod
-     * @param string          $route
-     * @param mixed           $handler
-     */
-    public function addRoute($httpMethod, $route, $handler);
-
-    /**
-     * Returns a listing of routes available.
-     *
-     * @return mixed
-     */
-    public function getRoutes();
 }