+++ src/ErrorHandler/ErrorHandlerInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rougin\Slytherin\ErrorHandler;
+
+/**
+ * Error Handler Interface
+ *
+ * An interface for handling third party debuggers.
+ * NOTE: To be removed in v1.0.0
+ *
+ * @package Slytherin
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+interface ErrorHandlerInterface extends \Rougin\Slytherin\Debug\DebuggerInterface
+{
+}