+++ src/IoC/Auryn/Container.php
@@ -7,49 +7,13 @@ namespace Rougin\Slytherin\IoC\Auryn;
  *
  * A simple implementation of a container that is built on top of
  * Daniel Lowrey's Auryn Inversion Of Control (IoC) Dependency Injector.
+ * NOTE: To be removed in v1.0.0
  *
  * https://github.com/rdlowrey/auryn
  *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class Container extends \Rougin\Slytherin\IoC\BaseContainer implements \Interop\Container\ContainerInterface
+class Container extends \Rougin\Slytherin\Container\Auryn\Container
 {
-    /**
-     * @var \Auryn\Injector
-     */
-    public $injector;
-
-    /**
-     * @param \Auryn\Injector $injector
-     */
-    public function __construct(\Auryn\Injector $injector)
-    {
-        $this->injector = $injector;
-    }
-
-    /**
-     * Adds a new instance to the container.
-     *
-     * @param string $id
-     * @param mixed  $concrete
-     */
-    public function add($id, $concrete = null)
-    {
-        if ($concrete && ! is_array($concrete)) {
-            $this->instances[$id] = $concrete;
-
-            return $this;
-        }
-
-        $arguments = array();
-
-        if (is_array($concrete)) {
-            $arguments = $concrete;
-        }
-
-        $this->instances[$id] = $this->injector->make($id, $arguments);
-
-        return $this;
-    }
 }